requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Embeddings & NLP
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Graph & Clustering
igraph>=0.10.0
leidenalg>=0.9.0

# Dimensionality Reduction & Visualization
umap-learn>=0.5.3
matplotlib>=3.5.0
seaborn>=0.12.0

# Data Processing
scipy>=1.7.0

# Configuration
pyyaml>=6.0

# Development Tools
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# Logging & Progress
tqdm>=4.64.0
rich>=13.0.0